chore: move zklend.ts to modules from src
